Added category-based loop and search bar
The loop now shows all posts except "Vortragsankündigungen". These will be shown on top of the loop if they are still in the future. As soon as the date and time are reached, they will not be displayed anymore

I also fixed a display problem when the excerpt text of a post was too long

I also designed the search bar according to the design